Title: 숫자 사이의 배수, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 231227/숫자 사이의 배수/multiplication-between-numbers.cpp
@@ -5,7 +5,7 @@ int main() {
     cin>>a>>b;
 
     int sum =0;
-    int average =0;
+    double average =0;
     int counter=0;
 
     for(int i=a; i<=b; i++)
